pyproject.toml
[tool.poetry]
name = "data-science"
version = "0.1"
description = "Data Science and Data QnA Multi-Agent"
authors = ["Meltem Subasioglu <msubasioglu@google.com>"]
license = "Apache License 2.0"
readme = "README.md"
package-mode = true 


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
google-adk = "^0.5.0"
immutabledict = "^4.2.1"
sqlglot = "^26.10.1"
db-dtypes = "^1.4.2"
regex = "^2024.11.6"
tabulate = "^0.9.0"
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}
absl-py = "^2.2.2"
litellm = "^1.30.7"
flask = "^2.0.1"
flask-socketio = "^5.1.1"
python-socketio = {extras = ["asyncio"], version = "^5.13.0"}
python-engineio = {extras = ["asyncio"], version = "^4.12.2"}
gevent = "^23.9.1"
gevent-websocket = "^0.10.1"
flask-cors = "^3.0.10"
hypercorn = "^0.14.3"
pandas = "^2.0.0"
numpy = "^1.26.4"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
plotly = "^5.18.0"
ydata-profiling = "^4.6.0"
dtale = "^3.12.0"
autoviz = "^0.1.0"
google-generativeai = "^0.8.5"
kaleido = "0.2.1"
google-cloud-texttospeech = "2.15.0"
google-cloud-vision = "^3.10.2"
google-cloud-bigquery = "^3.34.0"
google-cloud-speech = "^2.33.0"
aiofiles = "^24.1.0"
deprecated = "^1.2.14"



[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"


[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/"]
log_level="DEBUG"
log_cli=true
log_auto_indent=true
log_cli_date_format="%Y-%m-%d %H:%M:%S"
log_cli_format ="[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    #"error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    #'ignore:function ham\(\) is deprecated:DeprecationWarning',
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

